# api/openapi.yaml
openapi: 3.1.0
info:
  title: Online Sales WeChat Mini-Program API
  description: The API documentation for the Plan3 Online Sales WeChat Mini-Program project.
  version: 1.0.0
servers:
  - url: /api
    description: Mock Service Worker
paths:
...
  /orders/calculate:
    post:
      summary: Calculate order total
      tags:
        - Orders
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCalculationRequest"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCalculationResponse"
  /orders:
    post:
      summary: Create a new order
      tags:
        - Orders
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      summary: Get a list of orders
      tags:
        - Orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
  /orders/{id}:
    get:
      summary: Get order details
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /orders/{id}/cancel:
    patch:
      summary: Cancel an order
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    OrderCalculationRequest:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
              price:
                type: number
        couponCode:
          type: string
        addressId:
          type: string
    OrderCalculationResponse:
      type: object
      properties:
        subtotal:
          type: number
        shipping:
          type: number
        discount:
          type: number
        total:
          type: number
        couponDiscount:
          type: number
        shippingMethod:
          type: string
    CreateOrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
              price:
                type: number
              selectedVariants:
                type: object
        address:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            province:
              type: string
            city:
              type: string
            district:
              type: string
            detail:
              type: string
        couponCode:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled, refunded]
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              name:
                type: string
              image:
                type: string
              quantity:
                type: integer
              price:
                type: number
              selectedVariants:
                type: object
        address:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            province:
              type: string
            city:
              type: string
            district:
              type: string
            detail:
              type: string
        payment:
          type: object
          properties:
            method:
              type: string
            amount:
              type: number
            paidAt:
              type: string
              format: date-time
        shipping:
          type: object
          properties:
            method:
              type: string
            fee:
              type: number
            trackingNumber:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time